Some more InteliiJ lead improvements, this time taking advantage of all the most modern Java syntax up to Java 1.8
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@20605 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/PastaThrallData.java
@@ -262,7 +262,7 @@ public class PastaThrallData
 
 	public static int dataToId( Object[] data )
 	{
-		return data == null ? 0 : ((Integer)data[ 1 ]).intValue();
+		return data == null ? 0 : ( Integer ) data[ 1 ];
 	}
 
 	public static String dataToSetting( Object[] data )
@@ -272,7 +272,7 @@ public class PastaThrallData
 
 	public static int dataToSkillId( Object[] data )
 	{
-		return data == null ? 0 : ((Integer)data[ 3 ]).intValue();
+		return data == null ? 0 : ( Integer ) data[ 3 ];
 	}
 
 	public static Pattern dataToPattern1( Object[] data )
